+++ api/v1/class/notifications/notifications.arbeit.inc.php
@@ -40,6 +40,7 @@ namespace Arbeitszeit{
         }
 
         public function get_notifications_edit_html(){
+            if(!$this->nodes()->checkNode("notifications.inc", "get_notifications_edit_html")) return "<p>{$this->i18n["no_entries"]}</p>";
             Exceptions::error_rep("[NOTIFICATIONS] Getting notifications edit html...");
             $sql = "SELECT * FROM `kalender` ORDER BY id DESC;";
             $res = $this->db->sendQuery($sql);
@@ -70,6 +71,7 @@ namespace Arbeitszeit{
         }
 
         public function get_all_notifications(){
+            if(!$this->nodes()->checkNode("notifications.inc", "get_all_notifications")) return null;
             Exceptions::error_rep("[NOTIFICATIONS] Getting all notifications...");
             $sql = "SELECT * FROM `kalender` ORDER BY id DESC;";
             $res = $this->db->sendQuery($sql);
@@ -99,6 +101,7 @@ namespace Arbeitszeit{
         }
 
         public function get_notifications_html(){
+            if(!$this->nodes()->checkNode("notifications.inc", "get_notifications_html")) return null;
             Exceptions::error_rep("[NOTIFICATIONS] Getting notifications html...");
             $base_url = Arbeitszeit::get_app_ini()["general"]["base_url"];
             $sql = "SELECT * FROM `kalender` ORDER BY id DESC;";
@@ -137,6 +140,7 @@ namespace Arbeitszeit{
          * 
          */
         public function get_notifications_entry($id){
+            if(!$this->nodes()->checkNode("notifications.inc", "get_notifications_entry")) return null;
             Exceptions::error_rep("[NOTIFICATIONS] Getting notifications entry with ID '$id'...");
             $sql = "SELECT * FROM `kalender` WHERE id = ?";
             $res = $this->db->sendQuery($sql);
@@ -171,6 +175,7 @@ namespace Arbeitszeit{
          * @param bool|array Returns "true" on success and an error array on failure
          */
         public function create_notifications_entry($time, $date, $location, $comment){
+            if(!$this->nodes()->checkNode("notifications.inc", "create_notifications_entry")) return false;
             Exceptions::error_rep("[NOTIFICATIONS] Creating new notifications entry...");
             $sql = "INSERT INTO `kalender` (`datum`, `uhrzeit`, `ort`, `notiz`) VALUES (?, ?, ?, ?)";
             $res = $this->db->sendQuery($sql)->execute([$date, $time, $location, $comment]);
@@ -199,6 +204,7 @@ namespace Arbeitszeit{
          * @Note All parameters are required, if not set, the function will return an error array 
          */
         public function edit_notifications_entry($id, $time, $date, $location, $comment){
+            if(!$this->nodes()->checkNode("notifications.inc", "edit_notifications_entry")) return false;
             Exceptions::error_rep("[NOTIFICATIONS] Editing notifications entry with ID '$id'...");
             $sql = "UPDATE `kalender` SET `datum` = ?, `uhrzeit` = ?, `ort` = ?, `notiz` = ? WHERE id = ?;";
             $res = $this->db->sendQuery($sql)->execute([$date, $time, $location, $comment, $id]);
@@ -223,6 +229,7 @@ namespace Arbeitszeit{
          * @return bool|array Returns "true" on success and an error array on failure
          */
         public function delete_notifications_entry($id){
+            if(!$this->nodes()->checkNode("notifications.inc", "delete_notifications_entry")) return false;
             Exceptions::error_rep("[NOTIFICATIONS] Deleting notifications entry with ID '$id'...");
             $sql = "DELETE FROM `kalender` WHERE id = ?;";
             $res = $this->db->sendQuery($sql)->execute([$id]);
